Move sidebar stuff into its own folder
