+++ src/components/recording/sidebar/Brush.js
