+++ src/components/recording/sidebar/ClearButton.js
