+++ src/components/recording/sidebar/ClearButtonContainer.js
@@ -2,7 +2,7 @@ import { connect } from 'react-redux'
 
 import ClearButton from './ClearButton'
 
-import { persistCurrentEvent } from '../../actions/events'
+import { persistCurrentEvent } from '../../../actions/events'
 
 function mapStateToProps (state, ownProps) {
   return {}
