+++ src/components/recording/sidebar/DrawingSidebar.js
