+++ src/components/recording/sidebar/UndoButtonContainer.js
@@ -2,7 +2,7 @@ import { connect } from 'react-redux'
 
 import UndoButton from './UndoButton'
 
-import { persistCurrentEvent } from '../../actions/events'
+import { persistCurrentEvent } from '../../../actions/events'
 
 function mapStateToProps (state, ownProps) {
   const disabled = state.recording.undoHistory.length === 0
